Add IncludedProjects component
